Support multiple ooxml packages in the tree view
